+++ operatingSystem/IThread.h
@@ -6,12 +6,12 @@ public:
 	virtual ~IThread() = default;
 
 	void start();
-	void startQuantum(int timeQuantum);
+	void startQuantum(long long timeQuantum);
 	static void sleep(int milliseconds);
 	//virtual void stop() = 0;
 
 protected:
 	virtual void run() = 0;
-	virtual void runQuantum(int timeQuantum) = 0;
+	virtual void runQuantum(long long timeQuantum) = 0;
 };
