+++ operatingSystem/RRScheduler.h
@@ -11,12 +11,12 @@ class RRScheduler : public AScheduler, public IThread
 {
 	private:
 		std::queue<std::shared_ptr<Process>> processQueue;  // Queue of processes
-		int timeQuantum;  // quantum in CPU cycles
-		int delayPerExec; // delay between instruction executions
+		long long timeQuantum;  // quantum in CPU cycles
+		long long delayPerExec; // delay between instruction executions
 
 	public:
 		// Constructor
-		RRScheduler(int quantum, int delayExec, int nCores);
+		RRScheduler(long long quantum, long long delayExec, int nCores);
 
 		// Instantiates core list based on given number of cores
 		void instantiateCoreList();
@@ -24,8 +24,8 @@ class RRScheduler : public AScheduler, public IThread
 		// Runs the actual scheduler
 		//void addProcessRR(std::shared_ptr<Process> process);  // add process to RR queue
 		void run() override;
-		void runQuantum(int timeQuantum) override;
+		void runQuantum(long long timeQuantum) override;
 		void init() override;
 		void execute() override;
-		void executeQuantum(int timeQuantum) override;
+		void executeQuantum(long long timeQuantum) override;
 };
